Driver update and child classes
Updated the base driver to set configuration, output all 8 channels as a list of floats. I also created two child classes for the MCC 7018 and the SuperLogics 8017. I have updated the creep procedure to reflect this, but need to test out. There are small differences in some of the commands between the 7018 and 8017 when you look at the full library. We should fully test out each instrument file for each instrument to make sure our current implementation is fully compatible. One command we are missing that may be useful is actually updating the CJC value. We have a get command but not a set command. I need help with that.